requirements.txt
absl-py==2.1.0
accelerate==0.31.0
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
clip-interrogator==0.6.0
faiss-cpu==1.8.0
filelock==3.15.3
flatbuffers==24.3.25
fsspec==2024.6.0
ftfy==6.2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.24.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pillow==10.3.0
pip-chill==1.0.3
protobuf==4.25.3
psutil==6.0.0
Pygments==2.18.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
six==1.16.0
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-text==2.16.1
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
vectordb2==0.1.9
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
